service-workers/service-worker/navigationpreload.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>ServiceWorker: navigator.serviceWorker.navigationPreload</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
promise_test(async t => {
    const SCRIPT = 'resources/empty-worker.js';
    const SCOPE = 'resources/navigationpreload';
    let registration =
        await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
    let np = await registration.navigationPreload;
    assert_true(np instanceof NavigationPreloadManager, 'navigationPreload ' +
                'attribute return NavigationPreloadManager object');
}, "The navigationPreload attribute must return service worker " +
    "registration's NavigationPreloadManager object.");
</script>
